fix migration

+++ database/migrations/2023_04_24_000512_alter_change_relationship_in_order_details_table.php
@@ -6,6 +6,8 @@ use Illuminate\Support\Facades\Schema;
 
 class AlterChangeRelationshipInOrderDetailsTable extends Migration
 {
+    private $table_name = 'order_details';
+    private $column = 'order_id';
     /**
      * Run the migrations.
      *
@@ -13,10 +15,15 @@ class AlterChangeRelationshipInOrderDetailsTable extends Migration
      */
     public function up()
     {
-        Schema::table('order_details', function (Blueprint $table) {
-            $table->dropColumn('id');
-            $table->primary(['order_id', 'product_id']);
+        Schema::table($this->table_name, function (Blueprint $table) {
+            if (Schema::hasColumn($this->table_name, $this->column)) {
+                Schema::table($this->table_name, function (Blueprint $table) {
+                    $table->dropColumn('id');
+                    $table->primary([$this->column, 'product_id']);
+                });
+            }
         });
+
     }
 
     /**
